requirements.txt
Django>=5.1.15
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
